added tags page

+++ src/main/resources/static/js/panel.js
@@ -1,9 +1,25 @@
-$(document).on("click", ".user-delete-button", function () {
-    var username = $(this).data('username');
-    $(".modal-body #usernameHolder").text(username);
+$(document).on("click", ".delete-button", function () {
+    var deleteType  = $(this).data('delete-type');
+    var dataName = null;
+    var deleteUrl = null;
+    switch (deleteType) {
+        case 'username':
+            dataName = $(this).data('username');
+            deleteUrl = '/api/users/'+dataName;
+        break;
+        case 'tag':
+            dataName = $(this).data('tagname');
+            deleteUrl = '/api/tags/'+dataName;
+        break;
+    }
+
+    if(dataName == null)
+        return;
+
+    $(".modal-body #dataIdHolder").text(dataName);
     $('#deleteModal').modal('show').one('click', '#delete', function (e) {
         $.ajax({
-            url: '/api/users/'+username+'/delete',
+            url: deleteUrl,
             type: 'DELETE',
             success: function(result) {
                 location.reload();
